修改: msgmeta的id号使用int

+++ comm/msg.go
@@ -25,13 +25,13 @@ type SessionCloseNotify struct {
 }
 
 type RemoteCallREQ struct {
-	MsgID  uint32
+	MsgID  uint16
 	Data   []byte
 	CallID int64
 }
 
 type RemoteCallACK struct {
-	MsgID  uint32
+	MsgID  uint16
 	Data   []byte
 	CallID int64
 }
@@ -44,10 +44,10 @@ func (self *SessionCloseNotify) String() string  { return fmt.Sprintf("%+v", *se
 func (self *RemoteCallREQ) String() string       { return fmt.Sprintf("%+v", *self) }
 func (self *RemoteCallACK) String() string       { return fmt.Sprintf("%+v", *self) }
 
-func (self *RemoteCallREQ) GetMsgID() uint32   { return self.MsgID }
+func (self *RemoteCallREQ) GetMsgID() uint16   { return self.MsgID }
 func (self *RemoteCallREQ) GetMsgData() []byte { return self.Data }
 func (self *RemoteCallREQ) GetCallID() int64   { return self.CallID }
-func (self *RemoteCallACK) GetMsgID() uint32   { return self.MsgID }
+func (self *RemoteCallACK) GetMsgID() uint16   { return self.MsgID }
 func (self *RemoteCallACK) GetMsgData() []byte { return self.Data }
 func (self *RemoteCallACK) GetCallID() int64   { return self.CallID }
